add admin functionality

+++ pages/api/products/[id].js
@@ -5,7 +5,10 @@ const handler = async (req, res) => {
   const {
     method,
     query: { id },
+    cookies,
   } = req
+
+  const token = cookies.token
 
   dbConnect()
 
@@ -19,6 +22,10 @@ const handler = async (req, res) => {
   }
 
   if (method === 'PUT') {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json('Not authenticated!')
+    }
+
     try {
       const product = await Product.findByIdAndUpdate(id, req.body, {
         new: true,
@@ -30,6 +37,10 @@ const handler = async (req, res) => {
   }
 
   if (method === 'DELETE') {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json('Not authenticated!')
+    }
+
     try {
       await Product.findByIdAndDelete(id)
       res.status(200).json('The product has been deleted!')
